Refactor the #invite and #exists actions out from the RolesController into separate mixins.
